fixed several bugs
method extractDay was throwing exceptions, so i deleted it and replaced everything using java Date, Calendar and SimpleDateFormat, which also take into account the locale of the system.  Fixed the action bar back button problem occuring when after selecting one match and pressing it, the app was returning to the first tab of the main activity, basically because it was recreated from scratch.