Update FragmentInfoMatch.java
bug fix: trying to make a toast on a null context resulted on crash. The crash was particularly easy to reproduce by turning off internet and joining a match: upon turning it on, getContext() would return null because the activity selectMatch had already been terminated when the onComplete() was called.